+++ iotdb-core/datanode/pom.xml
@@ -288,7 +288,7 @@
         </dependency>
         <dependency>
             <groupId>org.apache.iotdb</groupId>
-            <artifactId>micrometer-metrics</artifactId>
+            <artifactId>metrics-core</artifactId>
             <version>1.3.0-SNAPSHOT</version>
         </dependency>
         <dependency>
