+++ iotdb-core/metrics/core/src/main/resources/META-INF/services/org.apache.iotdb.metrics.AbstractMetricManager
@@ -15,4 +15,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-org.apache.iotdb.metrics.micrometer.MicrometerMetricManager
+org.apache.iotdb.metrics.micrometer.IoTDBMetricManager
